Fix more Roslynator findings
