+++ src/UnitTests/VirtualFileSystem.cs
@@ -100,14 +100,14 @@ namespace AggregateConfigBuildTask.Tests.Unit
         }
 
         /// <inheritdoc/>
-        public bool DirectoryExists(string path)
+        public bool DirectoryExists(string directoryPath)
         {
-            path = NormalizePath(path);
-            path = EnsureTrailingDirectorySeparator(path);
+            directoryPath = NormalizePath(directoryPath);
+            directoryPath = EnsureTrailingDirectorySeparator(directoryPath);
 
             foreach (var file in fileSystem.Keys)
             {
-                if (file.Equals(path, StringComparison))
+                if (file.Equals(directoryPath, StringComparison))
                 {
                     return true;
                 }
@@ -117,12 +117,12 @@ namespace AggregateConfigBuildTask.Tests.Unit
         }
 
         /// <inheritdoc/>
-        public void CreateDirectory(string path)
+        public void CreateDirectory(string directoryPath)
         {
-            path = NormalizePath(path);
-            path = EnsureTrailingDirectorySeparator(path);
+            directoryPath = NormalizePath(directoryPath);
+            directoryPath = EnsureTrailingDirectorySeparator(directoryPath);
 
-            fileSystem[path] = string.Empty;
+            fileSystem[directoryPath] = string.Empty;
         }
 
         /// <inheritdoc />
@@ -163,6 +163,7 @@ namespace AggregateConfigBuildTask.Tests.Unit
         /// <summary>
         /// Converts a file search pattern (with * and ?) to a regex pattern.
         /// </summary>
+        /// <param name="searchPattern">The wild card path expression to convert to a regex pattern.</param>
         private static string ConvertPatternToRegex(string searchPattern)
         {
             // Escape special regex characters except for * and ?
